requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic[email]==2.10.5
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic-settings==2.7.0
google-generativeai
deep-translator
PyPDF2==3.0.1
python-docx==1.1.2
websockets==12.0
openai-whisper==20231117
gtts==2.5.0